refactor: remove unaffected stuff and collect rebuild modules in make
