+++ crates/rspack_core/src/compiler/make/mod.rs
@@ -9,9 +9,8 @@ use rustc_hash::FxHashSet as HashSet;
 
 use self::{cutout::Cutout, repair::repair};
 use crate::{
-  unaffected_cache::{Mutation, Mutations},
-  utils::FileCounter,
-  BuildDependency, Compilation, DependencyId, ModuleGraph, ModuleGraphPartial, ModuleIdentifier,
+  utils::FileCounter, BuildDependency, Compilation, DependencyId, ModuleGraph, ModuleGraphPartial,
+  ModuleIdentifier,
 };
 
 #[derive(Debug, Default)]
@@ -20,10 +19,10 @@ pub struct MakeArtifact {
   // should be reset when rebuild
   pub diagnostics: Vec<Diagnostic>,
   pub has_module_graph_change: bool,
-  pub mutations: Option<Mutations>,
 
   // data
   pub built_modules: IdentifierSet,
+  pub revoked_modules: IdentifierSet,
   pub make_failed_dependencies: HashSet<BuildDependency>,
   pub make_failed_module: IdentifierSet,
   pub module_graph_partial: ModuleGraphPartial,
@@ -50,16 +49,16 @@ impl MakeArtifact {
     &mut self.module_graph_partial
   }
 
-  pub fn take_mutations(&mut self) -> Option<Mutations> {
-    std::mem::take(&mut self.mutations)
-  }
-
   pub fn take_diagnostics(&mut self) -> Vec<Diagnostic> {
     std::mem::take(&mut self.diagnostics)
   }
 
   pub fn take_built_modules(&mut self) -> IdentifierSet {
     std::mem::take(&mut self.built_modules)
+  }
+
+  pub fn take_revoked_modules(&mut self) -> IdentifierSet {
+    std::mem::take(&mut self.revoked_modules)
   }
 
   fn revoke_module(&mut self, module_identifier: &ModuleIdentifier) -> Vec<BuildDependency> {
@@ -81,11 +80,7 @@ impl MakeArtifact {
         .build_dependencies
         .remove_batch_file(&build_info.build_dependencies);
     }
-    if let Some(mutations) = &mut self.mutations {
-      mutations.add(Mutation::ModuleRevoke {
-        module: *module_identifier,
-      });
-    }
+    self.revoked_modules.insert(*module_identifier);
     module_graph.revoke_module(module_identifier)
   }
 
@@ -143,11 +138,8 @@ pub fn make_module_graph(
 
   // reset temporary data
   artifact.built_modules = Default::default();
+  artifact.revoked_modules = Default::default();
   artifact.diagnostics = Default::default();
-  artifact.mutations = compilation
-    .incremental
-    .can_write_mutations()
-    .then(Default::default);
   artifact.has_module_graph_change = false;
 
   artifact = update_module_graph(compilation, artifact, params)?;
