+++ crates/rspack_core/src/compiler/make/repair/mod.rs
@@ -13,7 +13,6 @@ use super::MakeArtifact;
 use crate::{
   module_graph::{ModuleGraph, ModuleGraphPartial},
   old_cache::Cache as OldCache,
-  unaffected_cache::UnaffectedModulesCache,
   utils::task_loop::{run_task_loop, Task},
   BuildDependency, Compilation, CompilerOptions, DependencyType, Module, ModuleFactory,
   ModuleProfile, NormalModuleSource, ResolverFactory, SharedPluginDriver,
@@ -28,7 +27,6 @@ pub struct MakeTaskContext {
   pub resolver_factory: Arc<ResolverFactory>,
   pub loader_resolver_factory: Arc<ResolverFactory>,
   pub old_cache: Arc<OldCache>,
-  pub unaffected_modules_cache: Arc<UnaffectedModulesCache>,
   pub dependency_factories: HashMap<DependencyType, Arc<dyn ModuleFactory>>,
 
   pub artifact: MakeArtifact,
@@ -43,7 +41,6 @@ impl MakeTaskContext {
       resolver_factory: compilation.resolver_factory.clone(),
       loader_resolver_factory: compilation.loader_resolver_factory.clone(),
       old_cache: compilation.old_cache.clone(),
-      unaffected_modules_cache: compilation.unaffected_modules_cache.clone(),
       dependency_factories: compilation.dependency_factories.clone(),
       fs: compilation.input_filesystem.clone(),
       artifact,
@@ -70,7 +67,6 @@ impl MakeTaskContext {
       self.loader_resolver_factory.clone(),
       None,
       self.old_cache.clone(),
-      self.unaffected_modules_cache.clone(),
       None,
       Default::default(),
       Default::default(),
