+++ crates/rspack_core/src/compiler/module_executor/mod.rs
@@ -21,6 +21,7 @@ use self::{
   overwrite::OverwriteTask,
 };
 use super::make::{repair::MakeTaskContext, update_module_graph, MakeArtifact, MakeParam};
+use crate::incremental::Mutation;
 use crate::{
   task_loop::run_task_loop_with_event, Compilation, CompilationAsset, Context, Dependency,
   DependencyId, LoaderImportDependency, PublicPath,
@@ -59,11 +60,9 @@ impl ModuleExecutor {
       let modules = std::mem::take(&mut make_artifact.make_failed_module);
       params.push(MakeParam::ForceBuildModules(modules));
     }
+    make_artifact.built_modules = Default::default();
+    make_artifact.revoked_modules = Default::default();
     make_artifact.diagnostics = Default::default();
-    make_artifact.mutations = compilation
-      .incremental
-      .can_write_mutations()
-      .then(Default::default);
     make_artifact.has_module_graph_change = false;
 
     make_artifact = update_module_graph(compilation, make_artifact, params).unwrap_or_default();
@@ -138,15 +137,21 @@ impl ModuleExecutor {
     let diagnostics = self.make_artifact.take_diagnostics();
     compilation.extend_diagnostics(diagnostics);
 
-    if let Some(mutations) = compilation.incremental.mutations_write()
-      && let Some(make_mutations) = self.make_artifact.take_mutations()
-    {
-      mutations.extend(make_mutations);
-    }
-
     let built_modules = self.make_artifact.take_built_modules();
+    if let Some(mutations) = compilation.incremental.mutations_write() {
+      for id in &built_modules {
+        mutations.add(Mutation::ModuleRevoke { module: *id });
+      }
+    }
     for id in built_modules {
       compilation.built_modules.insert(id);
+    }
+
+    let revoked_modules = self.make_artifact.take_revoked_modules();
+    if let Some(mutations) = compilation.incremental.mutations_write() {
+      for id in revoked_modules {
+        mutations.add(Mutation::ModuleRevoke { module: id });
+      }
     }
 
     let code_generated_modules = std::mem::take(&mut self.code_generated_modules);
