+++ crates/rspack_core/src/lib.rs
@@ -11,7 +11,7 @@ mod cgm_hash_results;
 mod cgm_runtime_requirement_results;
 mod dependencies_block;
 pub mod diagnostics;
-pub mod unaffected_cache;
+pub mod incremental;
 pub use dependencies_block::{
   AsyncDependenciesBlock, AsyncDependenciesBlockIdentifier, DependenciesBlock,
 };
