+++ crates/rspack_plugin_javascript/src/plugin/infer_async_modules_plugin.rs
@@ -1,7 +1,7 @@
 use linked_hash_set::LinkedHashSet;
 use rspack_collections::IdentifierSet;
 use rspack_core::{
-  unaffected_cache::{IncrementalPasses, Mutation, Mutations},
+  incremental::{IncrementalPasses, Mutation, Mutations},
   ApplyContext, Compilation, CompilationFinishModules, CompilerOptions, DependencyType,
   ModuleGraph, ModuleIdentifier, Plugin, PluginContext,
 };
@@ -15,16 +15,11 @@ pub struct InferAsyncModulesPlugin;
 #[plugin_hook(CompilationFinishModules for InferAsyncModulesPlugin)]
 async fn finish_modules(&self, compilation: &mut Compilation) -> Result<()> {
   let module_graph = compilation.get_module_graph();
-  let modules: IdentifierSet = if compilation
+  let modules: IdentifierSet = if let Some(mutations) = compilation
     .incremental
-    .can_read_mutations(IncrementalPasses::INFER_ASYNC_MODULES)
+    .mutations_read(IncrementalPasses::INFER_ASYNC_MODULES)
   {
-    compilation
-      .unaffected_modules_cache
-      .get_affected_modules_with_module_graph()
-      .lock()
-      .expect("should lock")
-      .clone()
+    mutations.get_affected_modules_with_module_graph(&module_graph)
   } else {
     module_graph.modules().keys().copied().collect()
   };
